+++ resources/views/i18n/index.blade.php
@@ -1,10 +1,10 @@
 <div claas="i18n">
     <div>
         <ul>
-            <li><a href="{{ route('i18n.missing-all') }}">Missing All</a></li>
-            <li><a href="{{ route('i18n.missing-cs') }}">Missing CS</a></li>
-            <li><a href="{{ route('i18n.missing-en') }}">Missing EN</a></li>
-            <li><a href="{{ route('i18n.complete') }}">Complete</a></li>
+            <li><a href="{{ route('i18n.missing-all') }}">{{ __('ig-common::layouts.i18n.link.missing-all') }}</a></li>
+            <li><a href="{{ route('i18n.missing-cs') }}">{{ __('ig-common::layouts.i18n.link.missing-cs') }}</a></li>
+            <li><a href="{{ route('i18n.missing-en') }}">{{ __('ig-common::layouts.i18n.link.missing-en') }}</a></li>
+            <li><a href="{{ route('i18n.complete') }}">{{ __('ig-common::layouts.i18n.link.complete') }}</a></li>
         </ul>
     </div>
 </div>